bug fix
I fixed a bug where Figure 1 (the SCA graph) was created wrong because the indexing variable was made after the three datasets were merged. I also changed the squares of the dashboard at the bottom of the graph to dots as I felt those would be easier to read. Furthermore, I changed the alpha (transparency) of the dashboard graph to 1 to match the top part of the SCA graph.